# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Human Link",
    "version": "3.0.0",
    "description": "Généré depuis les routeurs Express et les schémas zod partagés (shared/src/index.ts). Authentification par cookie httpOnly (`hl_at`). Les écritures (POST, PUT, PATCH, DELETE) exigent aussi l’en-tête `X-CSRF-Token` recopiant le cookie `hl_csrf` (obtenu par GET /api/auth/csrf). Les dates sont au format AAAA-MM-JJ."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/calendar/leaves": {
      "get": {
        "tags": [
          "calendar"
        ],
        "operationId": "get_calendar_leaves",
        "summary": "Congés acceptés de l’équipe (manager) ou de tous (administrateur)",
        "description": "Rôles autorisés : admin, manager. Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "string"
                          },
                          "nom": {
                            "type": "string"
                          },
                          "prenom": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "start": {
                            "type": "string",
                            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                          },
                          "end": {
                            "type": "string",
                            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                          },
                          "raison": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "userId",
                          "nom",
                          "prenom",
                          "email",
                          "start",
                          "end",
                          "raison"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "admin",
          "manager"
        ]
      }
    },
...
    "/api/leaves": {
      "get": {
        "tags": [
          "leaves"
        ],
        "operationId": "get_leaves",
        "summary": "Demandes de congé à traiter",
        "description": "Un manager voit son équipe, un administrateur tout le monde. Rôles autorisés : admin, manager. Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "statut",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "en attente",
                "accepte",
                "refuse"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "user": {
                            "anyOf": [
                              {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "nom": {
                                    "type": "string"
                                  },
                                  "prenom": {
                                    "type": "string"
                                  },
                                  "email": {
                                    "type": "string"
                                  },
                                  "departement": {
                                    "type": "string"
                                  },
                                  "poste": {
                                    "type": "string"
                                  },
                                  "photoFileId": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "id",
                                  "nom",
                                  "prenom"
                                ],
                                "additionalProperties": false
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "raison": {
                            "type": "string"
                          },
                          "dateDebut": {
                            "type": "string",
                            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                          },
                          "dateFin": {
                            "type": "string",
                            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                          },
                          "nombreJours": {
                            "type": "number"
                          },
                          "description": {
                            "type": "string"
                          },
                          "statut": {
                            "type": "string",
                            "enum": [
                              "en attente",
                              "accepte",
                              "refuse"
                            ]
                          },
                          "statutManager": {
                            "type": "string",
                            "enum": [
                              "en attente",
                              "accepte",
                              "refuse"
                            ]
                          },
                          "statutAdmin": {
                            "type": "string",
                            "enum": [
                              "en attente",
                              "accepte",
                              "refuse"
                            ]
                          },
                          "motifRefus": {
                            "type": "string"
                          },
                          "attachmentFileId": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "userId",
                          "raison",
                          "dateDebut",
                          "dateFin",
                          "nombreJours",
                          "statut",
                          "statutManager",
                          "statutAdmin",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "admin",
          "manager"
        ]
      },
      "post": {
        "tags": [
          "leaves"
        ],
        "operationId": "post_leaves",
        "summary": "Déposer une demande de congé",
        "description": "Rôles autorisés : employe, manager. Authentification requise (cookie de session).",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "raison": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "dateDebut": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "dateFin": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "attachment": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "raison",
                  "dateDebut",
                  "dateFin"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "leave": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "userId": {
                          "type": "string"
                        },
                        "user": {
                          "anyOf": [
                            {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "nom": {
                                  "type": "string"
                                },
                                "prenom": {
                                  "type": "string"
                                },
                                "email": {
                                  "type": "string"
                                },
                                "departement": {
                                  "type": "string"
                                },
                                "poste": {
                                  "type": "string"
                                },
                                "photoFileId": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "id",
                                "nom",
                                "prenom"
                              ],
                              "additionalProperties": false
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "raison": {
                          "type": "string"
                        },
                        "dateDebut": {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                        },
                        "dateFin": {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                        },
                        "nombreJours": {
                          "type": "number"
                        },
                        "description": {
                          "type": "string"
                        },
                        "statut": {
                          "type": "string",
                          "enum": [
                            "en attente",
                            "accepte",
                            "refuse"
                          ]
                        },
                        "statutManager": {
                          "type": "string",
                          "enum": [
                            "en attente",
                            "accepte",
                            "refuse"
                          ]
                        },
                        "statutAdmin": {
                          "type": "string",
                          "enum": [
                            "en attente",
                            "accepte",
                            "refuse"
                          ]
                        },
                        "motifRefus": {
                          "type": "string"
                        },
                        "attachmentFileId": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "userId",
                        "raison",
                        "dateDebut",
                        "dateFin",
                        "nombreJours",
                        "statut",
                        "statutManager",
                        "statutAdmin",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "leave"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "employe",
          "manager"
        ]
      }
    },
    "/api/leaves/{id}": {
      "delete": {
        "tags": [
          "leaves"
        ],
        "operationId": "delete_leaves_id",
        "summary": "Retirer ma demande en attente",
        "description": "Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true
      }
    },
    "/api/leaves/{id}/attachment": {
      "get": {
        "tags": [
          "leaves"
        ],
        "operationId": "get_leaves_id_attachment",
        "summary": "Justificatif d’une demande",
        "description": "Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fichier",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hl_at"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}